requirements.txt
Django==3.1.6
django-extensions==3.1.1
django-polymorphic==3.0.0
